Update name of the repositories methods

+++ Repositories/CategoryRepository.cs
@@ -13,27 +13,39 @@ public class CategoryRepository : ICategoryRepository
         _context = context;
     }
 
-    public async Task<IEnumerable<Category>> GetAll() => await _context.Categories.ToListAsync();
+    public async Task<IEnumerable<Category>> GetAllAsync() => await _context.Categories.ToListAsync();
 
-    public async Task<Category?> GetById(int id) => await _context.Categories.FindAsync(id);
+    public async Task<Category?> GetByIdAsync(int id) => await _context.Categories.FindAsync(id);
 
-    public async Task Create(Category entity) => await _context.Categories.AddAsync(entity);
-
-    public void Update(Category entity) => _context.Categories.Update(entity);
-
-    public async Task Delete(int id)
+    public async Task<Category> CreateAsync(Category entity)
     {
-        var category = await GetById(id) ?? throw new KeyNotFoundException($"Category with ID {id} not found.");
-
-        _context.Categories.Remove(category);
+        await _context.Categories.AddAsync(entity);
+        return entity;
     }
 
-    public async Task SaveChanges() => await _context.SaveChangesAsync();
+    public Category Update(Category entity)
+    {
+        _context.Categories.Update(entity);
+        return entity;
+    }
 
-    public async Task<Category?> GetByName(string name) => await _context.Categories.FirstOrDefaultAsync(c => c.Name == name);
+    public async Task<bool> DeleteAsync(int id)
+    {
+        var category = await GetByIdAsync(id);
 
-    public async Task<int> GetUsageCount(int categoryId) => await _context.ToDos.CountAsync(t => t.CategoryId == categoryId);
+        if (category == null) return false;
 
-    public async Task<int> GetUsageCountByUser(int categoryId, int userId) =>
+        _context.Categories.Remove(category);
+
+        return true;
+    }
+
+    public async Task SaveChangesAsync() => await _context.SaveChangesAsync();
+
+    public async Task<Category?> GetByNameAsync(string name) => await _context.Categories.FirstOrDefaultAsync(c => c.Name == name);
+
+    public async Task<int> GetUsageCountAsync(int categoryId) => await _context.ToDos.CountAsync(t => t.CategoryId == categoryId);
+
+    public async Task<int> GetUsageCountByUserAsync(int categoryId, int userId) =>
         await _context.ToDos.CountAsync(t => t.CategoryId == categoryId && t.UserId == userId);
 }
